Update conversation script in ZSCRIPT.zc
This commit contains updates to various lines of dialogue in the ZSCRIPT.zc file. The conversations between Percival and Gasket have been rewritten to make the dialogue more emotional and authentic, adding a sense of dread and fear in anticipation of an impending end. Several silent cues have also been replaced with active lines.